data type cleanup

+++ Emulator/Agnus/Agnus.cpp
@@ -527,7 +527,7 @@ Agnus::computeDDFWindowOCS()
      *  7 | large   | medium  | -       || not handled         | DDF_OFF
      *  8 | large   | large   | -       || Empty               | DDF_OFF
      */
-    const struct { int interval; } table[9] = {
+    const struct { isize interval; } table[9] = {
         { DDF_EMPTY     }, // 0
         { DDF_18_STOP   }, // 1
         { DDF_18_D8     }, // 2
@@ -539,7 +539,7 @@ Agnus::computeDDFWindowOCS()
         { DDF_EMPTY     }  // 8
     };
 
-    int index = 3*strt + stop;
+    isize index = 3*strt + stop;
     switch (table[index].interval) {
 
         case DDF_EMPTY:
@@ -606,7 +606,7 @@ Agnus::computeDDFWindowECS()
      * 16 | large   | large   | DDF_OFF || Empty               | DDF_OFF
      * 17 | large   | large   | DDF_ON  || [0x18 ; 0xD8]       | DDF_ON
      */
-    const struct { int interval; DDFState state; } table[18] = {
+    const struct { isize interval; DDFState state; } table[18] = {
         { DDF_EMPTY ,    DDF_OFF }, // 0
         { DDF_EMPTY ,    DDF_OFF }, // 1
         { DDF_18_STOP ,  DDF_OFF }, // 2
@@ -627,7 +627,7 @@ Agnus::computeDDFWindowECS()
         { DDF_18_D8 ,    DDF_ON  }, // 17
     };
 
-    int index = 6*strt + 2*stop + (ddfState == DDF_ON);
+    isize index = 6*strt + 2*stop + (ddfState == DDF_ON);
     switch (table[index].interval) {
 
         case DDF_EMPTY:
